chore: completely migrate to eslint flat config
