--- .eslintignore
@@ -1,4 +1,0 @@
-node_modules
-dist
-temp
-coverage
