+++ package.json
@@ -13,7 +13,7 @@
     "size-esm-runtime": "node scripts/build.js vue -f esm-bundler-runtime",
     "size-esm": "node scripts/build.js runtime-dom runtime-core reactivity shared -f esm-bundler",
     "check": "tsc --incremental --noEmit",
-    "lint": "ESLINT_USE_FLAT_CONFIG=false eslint --cache --ext .js,.ts,.tsx . --ignore-path .gitignore",
+    "lint": "eslint --cache .",
     "format": "prettier --write --cache .",
     "format-check": "prettier --check --cache .",
     "test": "vitest",
@@ -51,7 +51,7 @@
       "prettier --write"
     ],
     "*.ts?(x)": [
-      "cross-env ESLINT_USE_FLAT_CONFIG=false eslint --fix",
+      "eslint --fix",
       "prettier --parser=typescript --write"
     ]
   },
@@ -72,8 +72,6 @@
     "@types/minimist": "^1.2.5",
     "@types/node": "^20.12.5",
     "@types/semver": "^7.5.8",
-    "@typescript-eslint/eslint-plugin": "^7.4.0",
-    "@typescript-eslint/parser": "^7.4.0",
     "@vitest/coverage-istanbul": "^1.4.0",
     "@vue/consolidate": "1.0.0",
     "conventional-changelog-cli": "^4.1.0",
@@ -113,12 +111,14 @@
     "tslib": "^2.6.2",
     "tsx": "^4.7.2",
     "typescript": "~5.4.5",
+    "typescript-eslint": "^7.6.0",
     "vite": "^5.2.7",
     "vitest": "^1.4.0"
   },
   "pnpm": {
     "peerDependencyRules": {
       "allowedVersions": {
+        "typescript-eslint>eslint": "^9.0.0",
         "@typescript-eslint/eslint-plugin>eslint": "^9.0.0",
         "@typescript-eslint/parser>eslint": "^9.0.0",
         "@typescript-eslint/type-utils>eslint": "^9.0.0",
